+++ backend/db.js
@@ -1,7 +1,10 @@
-const { Pool } = require('pg');
-require('dotenv').config();
+//Cambiado de require a import
+import { Pool } from 'pg';
+import dotenv from 'dotenv';
 
-const pool = new Pool({
+dotenv.config();
+
+export const pool = new Pool({
   host: process.env.DB_HOST,
   port: process.env.DB_PORT,
   user: process.env.DB_USER,
@@ -9,4 +12,3 @@ const pool = new Pool({
   database: process.env.DB_NAME,
 });
 
-module.exports = pool;
