+++ backend/package.json
@@ -7,10 +7,14 @@
     "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
   ],
   "license": "MIT",
+  "type": "module",
   "repository": "jshttp/accepts",
   "dependencies": {
+    "dotenv": "^16.5.0",
+    "express": "^5.1.0",
     "mime-types": "~2.1.34",
-    "negotiator": "0.6.3"
+    "negotiator": "0.6.3",
+    "pg": "^8.16.0"
   },
   "devDependencies": {
     "deep-equal": "1.0.1",
